htdocs/simulcast.html

<html>
<head>
    <meta charset="UTF-8"/>
    <meta name="timeout" content="long"/>
    <title>WebRTC WPT demo</title>
    <script src="https://wpt.live/resources/testharness.js"></script>
    <script src="https://wpt.live/resources/testharnessreport.js"></script>
    <script src="https://wpt.live/webrtc/third_party/sdp/sdp.js"></script>
    <script src="lib.js"></script>
</head>
<body>
<script>

async function waitForRid(ws, rid) {
    // hardcode rid extension id for now:
    // a=extmap:10 urn:ietf:params:rtp-hdrext:sdes:rtp-stream-id
    const ridExtensionId = 10;
    return new Promise((resolve) => {
        ws.addEventListener('message', function listener(message) {
            if (isRTCP(message.data)) {
                return;
            }
            const rtpData = new RTP(message.data);
            // We only want packets with the rid extension.
            if (!rtpData.headerExtensions) {
                return;
            }
            const ridExtension = rtpData.headerExtensions.find(ext => ext.id === ridExtensionId);
            if (!ridExtension) {
                return;
            }
            if (ridExtension.data.byteLength >= 1 && ridExtension.data.getUint8(0) === rid.charCodeAt(0)) {
                ws.removeEventListener('message', listener);
                resolve(message.data);
            }
        });
    });
}

/* times out since we have no BWE.
promise_test(async (t) => {
    const pc = new RTCPeerConnection();
    t.add_cleanup(() => pc.close());

    const stream = await navigator.mediaDevices.getUserMedia({
        audio: false,
        video: true
    });
    const rids = ['l', 'm', 'h'];
    pc.addTransceiver(stream.getVideoTracks()[0], {
        streams: [stream],
        sendEncodings: rids.map(rid => ({rid})),
    });

    const helper = new WebRTCHelper(pc);
    t.add_cleanup(() => helper.cleanup());

    const ws = await helper.connect();
    await helper.offer();

    await Promise.all(rids.map(rid => waitForRid(ws, rid)));
}, 'simulcast sends packet with RID extension');
*/

promise_test(async (t) => {
    const pc = new RTCPeerConnection();
    t.add_cleanup(() => pc.close());

    const stream = await navigator.mediaDevices.getUserMedia({
        audio: false,
        video: true
    });
    const rids = ['l', 'm', 'h'];
    pc.addTransceiver(stream.getVideoTracks()[0], {
        streams: [stream],
        sendEncodings: rids.map(rid => ({rid})),
    });
    const helper = new WebRTCHelper(pc);
    t.add_cleanup(() => helper.cleanup());

    const ws = await helper.connect();
    await helper.offer();

    const packet = new RTP(await waitForRid(ws, 'l'));
    const firstByte = packet.payload.getUint8(0);
    const secondByte = packet.payload.getUint8(1);
    // https://tools.ietf.org/html/rfc7741#section-4.2
    // X bit should be set.
    assert_true(!!(firstByte & 0x80), 'VP8 X bit is set');
    assert_true(!!(secondByte & 0x20), 'VP8 T bit is set');
}, 'Simulcast sends VP8 packet with temporal layers');
</script>
</body>
</html>
